Ensure that system apps are first in boot script
kernel, stdlib, and sasl (if included) should be the first three apps
in generated boot scripts.  This change ensures that this order is
always preserved.  For example, if sasl is included later in the boot
script it will miss progress reports for any applications started
beforehand.

I also include lager as a fourth item so that, if included, it will be
started as soon as is possible.  Perhaps this is best left to
individual app dependencies.